Split MCPServerMacro.swift into focused extension files
Refactor only — generated code (the macro's emitted strings) is
byte-for-byte unchanged; all 415 tests still pass.

Before: one 1307-line file, with a 477-line static expansion(...)
function (cyclomatic complexity 40) and a 1050-line struct body.
After: the main file is 200 lines, expansion(...) is ~48 lines (just
the dispatch), and the work is grouped by concern:

- MCPServerMacro+Helpers.swift — @MCPServer argument parsing, function
  discovery (collectTool/Resource/PromptFunctions, conformance checks).
- MCPServerMacro+DispatchGeneration.swift — tool/resource/prompt
  metadata properties + callTool/callResource/callPrompt switches +
  extension storage scaffolding (__mcpExtensionContributions, etc.).
- MCPServerMacro+ClientGeneration.swift — top-level Client type
  emission (makeClientType + doc-comment helpers).
- MCPServerMacro+ClientMethods.swift — per-method client body emitters
  for tools/resources/prompts.
- MCPServerMacro+ClientExpressions.swift — call-site signature helpers,
  template-variable parsing, small expression builders.
- MCPServerMacro+Conformances.swift — ExtensionMacro conformance for
  emitting MCPServer/Providing protocol conformances.
